fake

Module faker

Fake value generation for specific formats.

It is structured in a way such that the modules here describes the custom group for structs implementing Dummy to generate custom fake formats.

Within the module, there is multiple modules. With raw module providing a generic faker requiring a locale (faker::lorem::raw::Paragraph<L>) and the rest of the modules providing a localized faker (faker::lorem::en::Paragraph) as convenient functions.

Examples

use fake::Fake;
use fake::faker::lorem::en::*;

let words: Vec<String> = Words(3..5).fake();

Modules

address

administrative

automotive

barcode

boolean

company

creditcard

currency

filesystem

finance

internet

job

lorem

name

number

phone_number